Split getProfile to its own file
